bot.py
import asyncio
import os

from dotenv import load_dotenv
from aiogram import Bot, Dispatcher, Router, filters 
from aiogram.types import Message
from aiogram.filters import CommandStart, Command
from handlers import router

load_dotenv()
BOT_TOKEN = os.getenv("BOT_TOKEN")

async def main():
    bot = Bot(token=BOT_TOKEN) 
    dp = Dispatcher()
    dp.include_router(router)
    await dp.start_polling(bot)

if __name__ == '__main__':
    asyncio.run(main())